Implement room creation and joining flow
Introduce new static HTML pages for the landing page, room creation, and room display. Update Room and Occupant entities, services, and controllers to support the new flow. Refine DTOs for room creation and joining, and adjust room expiry listener logic.

+++ src/main/java/com/victor/EventDrop/rooms/Room.java
@@ -26,5 +26,5 @@ public class Room {
 
     //In Seconds
     @TimeToLive
-    private long ttl;
+    private double ttl;
 }
